requirements.txt
pillow
opencv-python
mss
pyautogui
numpy
python-xlib
pynput
flask_cors
flask
matplotlib